+++ apps/app_ninja/djninja/settings.py
@@ -2,7 +2,7 @@ import sys
 
 sys.path.insert(0, '../..')
 
-from common_django_settings import *  # noqa
+from apps.common_django_settings import *  # noqa
 
 ROOT_URLCONF = 'djninja.urls'
 WSGI_APPLICATION = 'djninja.wsgi.application'
