--- network_service.py
@@ -1,17 +1,0 @@
-import asyncio
-import os
-from sanic import Sanic
-from sanic.response import text
-
-
-app = Sanic("some_job")
-
-
-@app.route("<path:path>")
-async def test(request, path):
-    await asyncio.sleep(0.1)  # This is the work
-    return text('OK')
-
-
-if __name__ == "__main__":
-    app.run(host="0.0.0.0", port=int(os.environ.get("PORT", "8000")), workers=2)
